+++ app/src/main/java/com/example/taskapp/ui/auth/LoginFragment.kt
@@ -1,25 +1,25 @@
-package com.example.taskapp.ui
+package com.example.taskapp.ui.auth
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.taskapp.databinding.FragmentSplashBinding
+import com.example.taskapp.databinding.FragmentLoginBinding
 
 
-class SplashFragment : Fragment() {
-    private var _binding: FragmentSplashBinding? = null
+
+class LoginFragment : Fragment() {
+    private var _binding: FragmentLoginBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSplashBinding.inflate(inflater, container, false)
+        _binding = FragmentLoginBinding.inflate(inflater, container, false)
         return binding.root
     }
-
 
     override fun onDestroyView() {
         super.onDestroyView()